+++ rotkehlchen/exchanges/constants.py
@@ -1,6 +1,6 @@
 from rotkehlchen.types import EXTERNAL_EXCHANGES, Location
 
-EXCHANGES_WITH_PASSPHRASE = (Location.COINBASEPRO, Location.KUCOIN, Location.OKX)
+EXCHANGES_WITH_PASSPHRASE = (Location.KUCOIN, Location.OKX)
 EXCHANGES_WITHOUT_API_SECRET = (Location.BITPANDA,)
 
 # Exchanges for which we have supported modules
@@ -22,6 +22,6 @@ SUPPORTED_EXCHANGES = EXCHANGES_WITH_PASSPHRASE + EXCHANGES_WITHOUT_API_SECRET +
     Location.HTX,
 )
 
-DEAD_EXCHANGES = (Location.FTX, Location.FTXUS, Location.BITTREX)
+DEAD_EXCHANGES = (Location.FTX, Location.FTXUS, Location.BITTREX, Location.COINBASEPRO)
 # Exchanges for which we allow import via CSV
 ALL_SUPPORTED_EXCHANGES = SUPPORTED_EXCHANGES + EXTERNAL_EXCHANGES + DEAD_EXCHANGES